Cleanout "device.rs"
