+++ src/io/device.rs
@@ -15,21 +15,3 @@ pub trait Device {
     fn name(&self) -> String;
     fn id(&self) -> i32;
 }
-
-impl std::fmt::Debug for dyn Input {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "Sensor {{ name: {}, id: {}, kind: {}",
-            self.name(),
-            self.id(),
-            self.get_metadata().kind
-        )
-    }
-}
-
-/// Defines an interface for an input device that needs to be calibrated
-pub trait Calibrated {
-    /// Initiate the calibration procedures for a specific device instance.
-    fn calibrate(&self) -> bool;
-}
